+++ BossMod/Modules/Dawntrail/Extreme/Ex3QueenEternal/Ex3QueenEternalStates.cs
@@ -29,7 +29,7 @@ class Ex3QueenEternalStates : StateMachineBuilder
         P1RoyalDomain(id + 0xA0000, 8.2f);
         P1LegitimateForce(id + 0xB0000, 6.3f);
         P1RoyalDomain(id + 0xC0000, 3.1f);
-        Cast(id + 0xD0000, AID.AuthorityEternal, 3.2f, 10, "Enrage");
+        Cast(id + 0xD0000, (uint)AID.AuthorityEternal, 3.2f, 10, "Enrage");
     }
 
     private void Phase2(uint id)
@@ -39,12 +39,12 @@ class Ex3QueenEternalStates : StateMachineBuilder
         P2DimensionalDistortion(id + 0x20000, 7.2f);
         P2DyingMemory(id + 0x30000, 1.3f);
         P2RadicalShift(id + 0x40000, 11.4f);
-        ActorCast(id + 0x50000, _module.BossP2, AID.Preservation, 7.3f, 14, true, "Enrage");
+        ActorCast(id + 0x50000, _module.BossP2, (uint)AID.Preservation, 7.3f, 14, true, "Enrage");
     }
 
     private void P1Aethertithe(uint id, float delay)
     {
-        Cast(id, AID.Aethertithe, delay, 3);
+        Cast(id, (uint)AID.Aethertithe, delay, 3);
         ComponentCondition<Aethertithe>(id + 0x10, 5, comp => comp.AOE != null)
             .ActivateOnEnter<Aethertithe>();
         ComponentCondition<Aethertithe>(id + 0x11, 5.1f, comp => comp.NumCasts >= 1, "Cone 1")
@@ -62,7 +62,7 @@ class Ex3QueenEternalStates : StateMachineBuilder
 
     private void P1ProsecutionOfWar(uint id, float delay)
     {
-        Cast(id, AID.ProsecutionOfWar, delay, 5, "Tankbuster 1")
+        Cast(id, (uint)AID.ProsecutionOfWar, delay, 5, "Tankbuster 1")
             .ActivateOnEnter<ProsecutionOfWar>()
             .SetHint(StateMachine.StateHint.Tankbuster);
         ComponentCondition<ProsecutionOfWar>(id + 2, 3.2f, comp => comp.NumCasts > 1, "Tankbuster 2")
@@ -72,7 +72,7 @@ class Ex3QueenEternalStates : StateMachineBuilder
 
     private void P1DivideAndConquer(uint id, float delay)
     {
-        Cast(id, AID.DivideAndConquer, delay, 7.5f)
+        Cast(id, (uint)AID.DivideAndConquer, delay, 7.5f)
             .ActivateOnEnter<DivideAndConquerBait>()
             .ActivateOnEnter<DivideAndConquerAOE>();
         ComponentCondition<DivideAndConquerBait>(id + 0x10, 0.1f, comp => comp.NumCasts > 0, "Protean 1");
@@ -84,13 +84,13 @@ class Ex3QueenEternalStates : StateMachineBuilder
 
     private void P1RoyalDomain(uint id, float delay)
     {
-        Cast(id, AID.RoyalDomain, delay, 5, "Raidwide")
+        Cast(id, (uint)AID.RoyalDomain, delay, 5, "Raidwide")
             .SetHint(StateMachine.StateHint.Raidwide);
     }
 
     private void P1LegitimateForce(uint id, float delay)
     {
-        CastMulti(id, [AID.LegitimateForceFirstR, AID.LegitimateForceFirstL], delay, 8, "Side 1")
+        CastMulti(id, [(uint)AID.LegitimateForceFirstR, (uint)AID.LegitimateForceFirstL], delay, 8, "Side 1")
             .ActivateOnEnter<LegitimateForce>();
         ComponentCondition<LegitimateForce>(id + 0x10, 3.1f, comp => comp.NumCasts > 1, "Side 2")
             .DeactivateOnExit<LegitimateForce>();
@@ -98,10 +98,10 @@ class Ex3QueenEternalStates : StateMachineBuilder
 
     private void P1Coronation(uint id, float delay)
     {
-        Cast(id, AID.Coronation, delay, 3);
+        Cast(id, (uint)AID.Coronation, delay, 3);
         ComponentCondition<Coronation>(id + 0x10, 2.1f, comp => comp.Groups.Count > 0)
             .ActivateOnEnter<Coronation>();
-        Cast(id + 0x20, AID.AtomicRay, 1.1f, 3);
+        Cast(id + 0x20, (uint)AID.AtomicRay, 1.1f, 3);
         ComponentCondition<AtomicRay>(id + 0x30, 1.2f, comp => comp.Active)
             .ActivateOnEnter<AtomicRay>();
         ComponentCondition<Coronation>(id + 0x40, 4.9f, comp => comp.NumCasts > 0, "Coronation")
@@ -112,7 +112,7 @@ class Ex3QueenEternalStates : StateMachineBuilder
 
     private void P1AbsoluteAuthority(uint id, float delay)
     {
-        Cast(id, AID.AbsoluteAuthorityPuddles, delay, 10);
+        Cast(id, (uint)AID.AbsoluteAuthorityPuddles, delay, 10);
         ComponentCondition<AbsoluteAuthorityPuddles>(id + 0x10, 0.1f, comp => comp.Casters.Count > 0, "Puddles bait")
             .ActivateOnEnter<AbsoluteAuthorityPuddles>();
         ComponentCondition<AbsoluteAuthorityExpansionBoot>(id + 0x20, 10, comp => comp.NumCasts > 0, "Spread/stack")
@@ -129,15 +129,15 @@ class Ex3QueenEternalStates : StateMachineBuilder
 
     private void P1VirtualShiftWind(uint id, float delay)
     {
-        Cast(id, AID.VirtualShiftWind, delay, 5, "Raidwide (wind platform)")
+        Cast(id, (uint)AID.VirtualShiftWind, delay, 5, "Raidwide (wind platform)")
             .SetHint(StateMachine.StateHint.Raidwide);
-        Cast(id + 0x10, AID.LawsOfWind, 5.2f, 4);
+        Cast(id + 0x10, (uint)AID.LawsOfWind, 5.2f, 4);
         ComponentCondition<Aeroquell>(id + 0x20, 0.1f, comp => comp.Active)
             .ActivateOnEnter<Aeroquell>();
         ComponentCondition<Aeroquell>(id + 0x21, 5, comp => !comp.Active, "Party stacks")
             .DeactivateOnExit<Aeroquell>();
 
-        CastStartMulti(id + 0x100, [AID.LegitimateForceFirstR, AID.LegitimateForceFirstL], 5.1f)
+        CastStartMulti(id + 0x100, [(uint)AID.LegitimateForceFirstR, (uint)AID.LegitimateForceFirstL], 5.1f)
             .ActivateOnEnter<AeroquellTwister>(); // voidzones appear ~0.6s after stacks
         ComponentCondition<MissingLink>(id + 0x101, 0.8f, comp => comp.TethersAssigned, "Chains")
             .ActivateOnEnter<LegitimateForce>()
@@ -151,7 +151,7 @@ class Ex3QueenEternalStates : StateMachineBuilder
             .DeactivateOnExit<WindOfChange>()
             .DeactivateOnExit<MissingLink>();
 
-        Cast(id + 0x300, AID.WorldShatterP1, 3, 5, "Raidwide + platform end")
+        Cast(id + 0x300, (uint)AID.WorldShatterP1, 3, 5, "Raidwide + platform end")
             .SetHint(StateMachine.StateHint.Raidwide);
         ComponentCondition<AeroquellTwister>(id + 0x310, 2.6f, comp => !comp.Sources(Module).Any())
             .DeactivateOnExit<AeroquellTwister>();
@@ -161,9 +161,9 @@ class Ex3QueenEternalStates : StateMachineBuilder
 
     private void P1VirtualShiftEarth(uint id, float delay)
     {
-        Cast(id, AID.VirtualShiftEarth, delay, 5, "Raidwide (earth platform)")
+        Cast(id, (uint)AID.VirtualShiftEarth, delay, 5, "Raidwide (earth platform)")
             .SetHint(StateMachine.StateHint.Raidwide);
-        CastStart(id + 0x10, AID.LawsOfEarth, 5.2f)
+        CastStart(id + 0x10, (uint)AID.LawsOfEarth, 5.2f)
             .ActivateOnEnter<VirtualShiftEarth>();
         CastEnd(id + 0x11, 4);
 
@@ -172,7 +172,7 @@ class Ex3QueenEternalStates : StateMachineBuilder
             .ActivateOnEnter<LawsOfEarthBurst1>()
             .DeactivateOnExit<LawsOfEarthBurst>();
 
-        Cast(id + 0x200, AID.GravitationalEmpire, 5.2f, 7)
+        Cast(id + 0x200, (uint)AID.GravitationalEmpire, 5.2f, 7)
             .ActivateOnEnter<GravityPillar>()
             .ActivateOnEnter<GravityRay>()
             .ActivateOnEnter<LawsOfEarthBurst2>();
@@ -190,31 +190,31 @@ class Ex3QueenEternalStates : StateMachineBuilder
         ComponentCondition<MeteorImpact>(id + 0x311, 6.1f, comp => comp.NumCasts > 0, "Meteors 2")
             .DeactivateOnExit<MeteorImpact>();
 
-        Cast(id + 0x400, AID.WeightyBlow, 2, 5);
+        Cast(id + 0x400, (uint)AID.WeightyBlow, 2, 5);
         ComponentCondition<WeightyBlow>(id + 0x410, 0.1f, comp => comp.NumCasts > 0, "LOS 1");
         ComponentCondition<WeightyBlow>(id + 0x411, 3.1f, comp => comp.NumCasts > 2, "LOS 2");
         ComponentCondition<WeightyBlow>(id + 0x412, 3.1f, comp => comp.NumCasts > 4, "LOS 3");
         ComponentCondition<WeightyBlow>(id + 0x413, 3.1f, comp => comp.NumCasts > 6, "LOS 4")
             .DeactivateOnExit<WeightyBlow>();
 
-        Cast(id + 0x500, AID.WorldShatterP1, 0.7f, 5, "Raidwide + platform end")
+        Cast(id + 0x500, (uint)AID.WorldShatterP1, 0.7f, 5, "Raidwide + platform end")
             .DeactivateOnExit<VirtualShiftEarth>()
             .SetHint(StateMachine.StateHint.Raidwide);
     }
 
     private void P1VirtualShiftIce(uint id, float delay)
     {
-        Cast(id, AID.VirtualShiftIce, delay, 5, "Raidwide (ice platform)")
+        Cast(id, (uint)AID.VirtualShiftIce, delay, 5, "Raidwide (ice platform)")
             .ActivateOnEnter<VirtualShiftIce>()
             .SetHint(StateMachine.StateHint.Raidwide);
-        CastStart(id + 0x10, AID.LawsOfIce, 5.2f);
+        CastStart(id + 0x10, (uint)AID.LawsOfIce, 5.2f);
         CastEnd(id + 0x11, 4)
             .ActivateOnEnter<LawsOfIce>();
         ComponentCondition<LawsOfIce>(id + 0x12, 1, comp => comp.NumCasts > 0, "Move");
         ComponentCondition<Rush>(id + 0x100, 4.3f, comp => comp.Activation != default)
             .ActivateOnEnter<Rush>()
             .DeactivateOnExit<LawsOfIce>();
-        CastStartMulti(id + 0x101, [AID.LegitimateForceFirstR, AID.LegitimateForceFirstL], 11.9f);
+        CastStartMulti(id + 0x101, [(uint)AID.LegitimateForceFirstR, (uint)AID.LegitimateForceFirstL], 11.9f);
         ComponentCondition<Rush>(id + 0x102, 0.3f, comp => comp.NumCasts > 0, "Stretch tethers")
             .DeactivateOnExit<Rush>();
         CastEnd(id + 0x103, 7.7f, "Side 1")
@@ -222,7 +222,7 @@ class Ex3QueenEternalStates : StateMachineBuilder
         ComponentCondition<LegitimateForce>(id + 0x104, 3.1f, comp => comp.NumCasts > 1, "Side 2")
             .DeactivateOnExit<LegitimateForce>();
 
-        Cast(id + 0x200, AID.LawsOfIce, 6.1f, 4)
+        Cast(id + 0x200, (uint)AID.LawsOfIce, 6.1f, 4)
             .ActivateOnEnter<LawsOfIce>();
         ComponentCondition<LawsOfIce>(id + 0x202, 1, comp => comp.NumCasts > 0, "Move")
             .ActivateOnEnter<IceDart>()
@@ -235,14 +235,14 @@ class Ex3QueenEternalStates : StateMachineBuilder
             .DeactivateOnExit<IceDart>()
             .DeactivateOnExit<RaisedTribute>();
 
-        Cast(id + 0x300, AID.WorldShatterP1, 3.1f, 5, "Raidwide + platform end")
+        Cast(id + 0x300, (uint)AID.WorldShatterP1, 3.1f, 5, "Raidwide + platform end")
             .DeactivateOnExit<VirtualShiftIce>()
             .SetHint(StateMachine.StateHint.Raidwide);
     }
 
     private void P2Intermission(uint id, float delay)
     {
-        Cast(id, AID.AuthorityEternal, delay, 10);
+        Cast(id, (uint)AID.AuthorityEternal, delay, 10);
         Targetable(id + 0x10, false, 0.2f, "Boss disappears + Raidwide")
             .SetHint(StateMachine.StateHint.Raidwide);
         ActorTargetable(id + 0x20, _module.BossP2, true, 24.8f, "Boss appears")
@@ -251,18 +251,18 @@ class Ex3QueenEternalStates : StateMachineBuilder
 
     private void P2RadicalShift(uint id, float delay)
     {
-        ActorCast(id, _module.BossP2, AID.RadicalShift, delay, 11, true, "Raidwide (platform change)")
+        ActorCast(id, _module.BossP2, (uint)AID.RadicalShift, delay, 11, true, "Raidwide (platform change)")
             .ActivateOnEnter<RadicalShift>()
             .ActivateOnEnter<VirtualShiftIce>()
             .ActivateOnEnter<RadicalShiftAOE>()
             .SetHint(StateMachine.StateHint.Raidwide);
         ComponentCondition<RadicalShiftAOE>(id + 0x10, 5.2f, comp => comp.NumFinishedSpreads > 0, "Spread")
             .ResetComp<RadicalShiftAOE>();
-        ActorCast(id + 0x100, _module.BossP2, AID.RadicalShift, 3, 11, true, "Raidwide (platform change)")
+        ActorCast(id + 0x100, _module.BossP2, (uint)AID.RadicalShift, 3, 11, true, "Raidwide (platform change)")
             .SetHint(StateMachine.StateHint.Raidwide);
         ComponentCondition<RadicalShiftAOE>(id + 0x110, 5.2f, comp => comp.NumFinishedSpreads > 0, "Spread")
             .DeactivateOnExit<RadicalShiftAOE>();
-        ActorCast(id + 0x200, _module.BossP2, AID.WorldShatterP2, 3, 5, true, "Raidwide + platform end")
+        ActorCast(id + 0x200, _module.BossP2, (uint)AID.WorldShatterP2, 3, 5, true, "Raidwide + platform end")
             .DeactivateOnExit<RadicalShift>()
             .DeactivateOnExit<VirtualShiftIce>()
             .SetHint(StateMachine.StateHint.Raidwide);
@@ -270,11 +270,11 @@ class Ex3QueenEternalStates : StateMachineBuilder
 
     private void P2DimensionalDistortion(uint id, float delay)
     {
-        ActorCast(id, _module.BossP2, AID.DimensionalDistortion, delay, 4, true)
+        ActorCast(id, _module.BossP2, (uint)AID.DimensionalDistortion, delay, 4, true)
             .ActivateOnEnter<DimensionalDistortion>();
         ComponentCondition<DimensionalDistortion>(id + 0x10, 1, comp => comp.NumCasts > 0, "Exaflares start");
 
-        ActorCast(id + 0x100, _module.BossP2, AID.TyrannysGrasp, 5.2f, 5, true, "Front half cleave")
+        ActorCast(id + 0x100, _module.BossP2, (uint)AID.TyrannysGrasp, 5.2f, 5, true, "Front half cleave")
             .ActivateOnEnter<TyrannysGraspAOE>()
             .ActivateOnEnter<TyrannysGraspTowers>()
             .DeactivateOnExit<DimensionalDistortion>()
@@ -297,7 +297,7 @@ class Ex3QueenEternalStates : StateMachineBuilder
             .DeactivateOnExit<DyingMemoryLast>()
             .SetHint(StateMachine.StateHint.Raidwide);
 
-        ActorCastStart(id + 0x100, _module.BossP2, AID.RoyalBanishment, 3.1f, true)
+        ActorCastStart(id + 0x100, _module.BossP2, (uint)AID.RoyalBanishment, 3.1f, true)
             .ActivateOnEnter<RoyalBanishment>(); // icon appears right before cast start
         ActorCastEnd(id + 0x101, _module.BossP2, 5, true);
         ComponentCondition<RoyalBanishment>(id + 0x110, 0.8f, comp => comp.NumCasts > 0, "Line stack 1");
